+++ android/src/main/kotlin/com/shatsy/pinchzoomimage/PinchZoomImagePlugin.kt
@@ -1,0 +1,25 @@
+package com.shatsy.pinchzoomimage
+
+import io.flutter.plugin.common.MethodChannel
+import io.flutter.plugin.common.MethodChannel.MethodCallHandler
+import io.flutter.plugin.common.MethodChannel.Result
+import io.flutter.plugin.common.MethodCall
+import io.flutter.plugin.common.PluginRegistry.Registrar
+
+class PinchZoomImagePlugin(): MethodCallHandler {
+  companion object {
+    @JvmStatic
+    fun registerWith(registrar: Registrar): Unit {
+      val channel = MethodChannel(registrar.messenger(), "pinch_zoom_image")
+      channel.setMethodCallHandler(PinchZoomImagePlugin())
+    }
+  }
+
+  override fun onMethodCall(call: MethodCall, result: Result): Unit {
+    if (call.method.equals("getPlatformVersion")) {
+      result.success("Android ${android.os.Build.VERSION.RELEASE}")
+    } else {
+      result.notImplemented()
+    }
+  }
+}