+++ ios/Classes/PinchZoomImagePlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface PinchZoomImagePlugin : NSObject<FlutterPlugin>
+@end