+++ ios/Classes/SwiftPinchZoomImagePlugin.swift
@@ -1,0 +1,14 @@
+import Flutter
+import UIKit
+
+public class SwiftPinchZoomImagePlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {
+    let channel = FlutterMethodChannel(name: "pinch_zoom_image", binaryMessenger: registrar.messenger())
+    let instance = SwiftPinchZoomImagePlugin()
+    registrar.addMethodCallDelegate(instance, channel: channel)
+  }
+
+  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+    result("iOS " + UIDevice.current.systemVersion)
+  }
+}