+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.2.0] - Oct 10
+
+* Turn package into plugin
+* Dim and show status bar using native platform code
+
+
 ## [0.1.1] - Sep 27
 
 * First release